## Related Issue
- Issue #16 

## 작업 상세 내용
- 공유 오피스 별점 정보 불러오기

+++ app/src/main/java/com/gangwonhyuil/gangwonhyuil/data/response/officeRating/officeRatingResponse.kt
@@ -8,5 +8,5 @@ data class officeRatingResponse (
     @SerializedName("total_reviews")
     val numOfReviews: Int,
     @SerializedName("average_rating")
-    val rating: Int
+    val rating: Float
 )
